+++ sources/gookto_io/InputAgents.cpp
@@ -6,7 +6,7 @@
 #include <gookto_io/tinyapi/tinyxml.h>
 #include <gookto_io/tinyapi/tinystr.h>
 
-InputAgents::InputAgents(int id, int type, float dpt_time) :
+InputAgents::InputAgents(int id, int type, double dpt_time) :
 id(id), type(type), dpt_time(dpt_time)
 {};
